Authentication ready on backend - configured authentication - further refactoring of entities needed

+++ backend/src/main/java/com/deroahe/gallerybe/model/Image.java
@@ -22,7 +22,7 @@ public class Image {
     private String imageUrl;
 
     @Column(name = "IMAGE_UPLOADED_BY")
-    private int imageUploadedBy;
+    private Long imageUploadedBy;
 
     // todo upload-date
